Add InitGenerator class to generate the loader.
This class currently doesn't observe variables but ensures that the userscript
runs in page scope, rather than in its own sandbox.

The technique behind this is to manipulate the dom tree by inserting a new
script element with the userscript code and as soon as it is evaluated removing
it again.